+++ db/schema.sql
@@ -1,0 +1,2 @@
+DROP DATABASE IF EXISTS users_db;
+CREATE DATABASE users_db;
