--- develop/db/schema.sql
@@ -1,2 +1,0 @@
-DROP DATABASE IF EXISTS users_db;
-CREATE DATABASE users_db;
