+++ seeds/seeds.js
@@ -1,0 +1,25 @@
+//const sequelize = require('../config/connection');
+//const { User, Project } = require('../models');
+
+//const userData = require('./userData.json');
+//const projectData = require('./projectData.json');
+
+//const seedDatabase = async () => {
+  //await sequelize.sync({ force: true });
+
+  //const users = await User.bulkCreate(userData, {
+    //individualHooks: true,
+    //returning: true,
+  
+
+ // for (const  of projectData) {
+   // await Project.create({
+     // ...project,
+      //user_id: users[Math.floor(Math.random() * users.length)].id,
+
+  
+
+ // process.exit(0);
+
+
+// seedDatabase();
