+++ views/dashboard.handlebars
@@ -1,0 +1,55 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Dashboard</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css">
+    <link rel="stylesheet" href="./css/dashboard.css">
+</head>
+<body>
+    <div class="container">
+        <!--main content-->
+        <div class="row">
+            <div class="col exercise">
+                <h2>Exercise Generator</h2>
+                <div class="body-part">
+                    <button>Back</button>
+                    <button>Cardio</button>
+                    <button>Chest</button>
+                    <button>Lower Arms</button>
+                    <button>Lower Legs</button>
+                    <button>Neck</button>
+                    <button>Shoulders</button>
+                    <button>Upper Arms</button>
+                    <button>Upper Legs</button>
+                    <button>Waist</button>
+                </div>
+                <div class="one-exercise">
+                    <div class="exercise-header">
+                        <button class="back-btn">Back</button>
+                        <h3 class="exercise-name">Name</h3>
+                    </div>
+                    <p class="exercise-target">Target</p>
+                    <p class="exercise-equipment">Equipment</p>
+                    <p class="exercise-description">Description</p>
+                    <img src="" class="exercise-gif">
+                </div>
+            </div>
+            <div class="col routine">
+                <h2>Exercise Tracker</h2>
+                <div class="tracker">
+                    <!--div for graph to generate-->
+                    <div id="chart"></div>
+                    <button class="completed-btn">Completed Exercise!</button>
+            </div>
+        </div>
+    </div>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
+    {{!-- <script src="anime.min.js"></script> --}}
+    <script src="/js/api.js"></script>
+    <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
+    <script src="/js/graph.js"></script>
+    <script src="https://code.highcharts.com/highcharts.js"></script>
+</body>
+</html>
