+++ views/profile.handlebars
@@ -1,0 +1,129 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>User Profile</title>
+    <link rel="stylesheet" href="/assets/style.css">
+    <link rel='stylesheet' href='https://cdn.jsdelivr.net/npm/bootstrap@4.4.1/dist/css/bootstrap.min.css'>
+    <script src='https://cdn.jsdelivr.net/npm/bootstrap@4.4.1/dist/js/bootstrap.bundle.min.js'></script>
+</head>
+<body>
+     <!-- Breadcrumb navigation bar here-->
+     <div class="container bootstrap snippets bootdey">
+      <div class="panel-body inf-content">
+          <div class="row">
+              <div class="col-md-4">
+                  <img alt="" style="width:600px;" title="" class="img-circle img-thumbnail isTooltip" src="/css/images/blank-profile-picture-circle-png-2548-3819284841.png" data-original-title="Usuario"> 
+                  <ul title="Ratings" class="list-inline ratings text-center">
+                      <li><a href="#"><span class="glyphicon glyphicon-star"></span></a></li>
+                      <li><a href="#"><span class="glyphicon glyphicon-star"></span></a></li>
+                      <li><a href="#"><span class="glyphicon glyphicon-star"></span></a></li>
+                      <li><a href="#"><span class="glyphicon glyphicon-star"></span></a></li>
+                      <li><a href="#"><span class="glyphicon glyphicon-star"></span></a></li>
+                  </ul>
+              </div>
+              <div class="col-md-6">
+                  <strong>Information</strong><br>
+                  <div class="table-responsive">
+                  <table class="table table-user-information">
+                      <tbody>
+                          <tr>        
+                              <td>
+                                  <strong>
+                                      <span class="glyphicon glyphicon-asterisk text-primary"></span>
+                                      Identification                                               
+                                  </strong>
+                              </td>
+                              <td class="text-primary">
+                                  123456789     
+                              </td>
+                          </tr>
+                          <tr>    
+                              <td>
+                                  <strong>
+                                      <span class="glyphicon glyphicon-user  text-primary"></span>    
+                                      Name                                                
+                                  </strong>
+                              </td>
+                              <td class="text-primary">
+                                  Bootdey     
+                              </td>
+                          </tr>
+                          <tr>        
+                              <td>
+                                  <strong>
+                                      <span class="glyphicon glyphicon-cloud text-primary"></span>  
+                                      Lastname                                                
+                                  </strong>
+                              </td>
+                              <td class="text-primary">
+                                  Bootstrap  
+                              </td>
+                          </tr>
+      
+                          <tr>        
+                              <td>
+                                  <strong>
+                                      <span class="glyphicon glyphicon-bookmark text-primary"></span> 
+                                      Username                                                
+                                  </strong>
+                              </td>
+                              <td class="text-primary">
+                                  N/A
+                              </td>
+                          </tr>
+      
+      
+                          <tr>        
+                              <td>
+                                  <strong>
+                                      <span class="glyphicon glyphicon-eye-open text-primary"></span> 
+                                      Role                                                
+                                  </strong>
+                              </td>
+                              <td class="text-primary">
+                                  Placeholder
+                              </td>
+                          </tr>
+                          <tr>        
+                              <td>
+                                  <strong>
+                                      <span class="glyphicon glyphicon-envelope text-primary"></span> 
+                                      Email                                                
+                                  </strong>
+                              </td>
+                              <td class="text-primary">
+                                  Placeholder
+                              </td>
+                          </tr>
+                          <tr>        
+                              <td>
+                                  <strong>
+                                      <span class="glyphicon glyphicon-calendar text-primary"></span>
+                                      created                                                
+                                  </strong>
+                              </td>
+                              <td class="text-primary">
+                                  19 May 2024
+                              </td>
+                          </tr>
+                          <tr>        
+                              <td>
+                                  <strong>
+                                      <span class="glyphicon glyphicon-calendar text-primary"></span>
+                                      Modified                                                
+                                  </strong>
+                              </td>
+                              <td class="text-primary">
+                                   19 May 2024
+                              </td>
+                          </tr>                                    
+                      </tbody>
+                  </table>
+                </div>
+            </div>
+          </div>
+      </div>
+    </div>
+    <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
+</body>
+</html>
